+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to
   `2 + 1000 // len(train)` to `2 + 10000 // len(train)`. This is due to the
   fact that models tended to stop training too early.
 
+### Fixed
+- Enforcing at most one GPU, as training on multiple GPUs led to tensors being
+  allocated to different devices.
+
 
 ## [v1.1.0] - 2021-09-13
 ### Added